+++ src/components/modules/banners/contributions/ContributionsBannerCta.tsx
@@ -1,0 +1,66 @@
+import { ThemeProvider } from 'emotion-theming';
+import { buttonReaderRevenueBrandAlt } from '@guardian/src-button/themes';
+import { LinkButton } from '@guardian/src-button/index';
+import { SvgArrowRightStraight } from '@guardian/src-icons';
+import React from 'react';
+import { css } from '@emotion/core';
+import { from } from '@guardian/src-foundations/mq';
+
+const styles = {
+    ctaButton: css`
+        margin-bottom: 0.5rem;
+    `,
+    cta: css`
+        display: flex;
+        flex-direction: column;
+        align-items: flex-start;
+        ${from.tablet} {
+            width: auto;
+            justify-content: flex-start;
+            flex-direction: column;
+        }
+    `,
+    paymentMethods: css`
+        display: block;
+        max-height: 1.25rem;
+    `,
+};
+
+interface ContributionsBannerCtaProps {
+    onContributeClick: () => void;
+    ctaText: string;
+    ctaUrl: string;
+}
+
+export const ContributionsBannerCta: React.FC<ContributionsBannerCtaProps> = ({
+    onContributeClick,
+    ctaText,
+    ctaUrl,
+}: ContributionsBannerCtaProps) => {
+    return (
+        <div css={styles.cta}>
+            <ThemeProvider theme={buttonReaderRevenueBrandAlt}>
+                <LinkButton
+                    data-link-name="contributions-banner : cta"
+                    css={styles.ctaButton}
+                    priority="primary"
+                    size="small"
+                    icon={<SvgArrowRightStraight />}
+                    iconSide="right"
+                    nudgeIcon={true}
+                    onClick={onContributeClick}
+                    hideLabel={false}
+                    aria-label="Contribute"
+                    href={ctaUrl}
+                >
+                    {ctaText}
+                </LinkButton>
+            </ThemeProvider>
+            <img
+                src="https://assets.guim.co.uk/images/acquisitions/2db3a266287f452355b68d4240df8087/payment-methods.png"
+                alt="Accepted payment methods: Visa, Mastercard, American Express and PayPal"
+                css={styles.paymentMethods}
+            />
+        </div>
+    );
+};
