+++ src/components/modules/banners/contributions/ContributionsBannerMobile.tsx
@@ -4,10 +4,8 @@ import { css } from '@emotion/core';
 import { neutral } from '@guardian/src-foundations';
 import { from } from '@guardian/src-foundations/mq';
 import { headline } from '@guardian/src-foundations/typography';
-import {
-    ContributionsBannerCloseButton,
-    ContributionsBannerCta,
-} from './ContributionsBannerButtons';
+import { ContributionsBannerCta } from './ContributionsBannerCta';
+import { ContributionsBannerCloseButton } from "./ContributionsBannerCloseButton";
 
 const mobileStyles = {
     container: css`
